refactor(hooks): extract useIdeLayoutState from ide-layout
