+++ apps/web/src/components/ide/layout/ide-layout.tsx
@@ -2,9 +2,18 @@
 
 import { usePathname } from "@i18n/routing";
 import { Sheet, SheetContent, TooltipProvider } from "@portfolio/ui";
+import dynamic from "next/dynamic";
 import { parseAsBoolean, useQueryState } from "nuqs";
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import { CommandPalette } from "@/components/ide/command/command-palette";
+import { useMemo, useRef } from "react";
+
+const CommandPalette = dynamic(
+  () =>
+    import("@/components/ide/command/command-palette").then((m) => ({
+      default: m.CommandPalette,
+    })),
+  { ssr: false }
+);
+
 import { ActivityBar } from "@/components/ide/layout/activity-bar";
 import { IdeEditorArea } from "@/components/ide/layout/ide-editor-area";
 import { IdeLayoutEmbed } from "@/components/ide/layout/ide-layout-embed";
@@ -12,20 +21,14 @@ import { MobileActivityBar } from "@/components/ide/layout/mobile-activity-bar";
 import { MobileMenu } from "@/components/ide/layout/mobile-menu";
 import { StatusBar } from "@/components/ide/layout/status-bar";
 import { TitleBar } from "@/components/ide/layout/title-bar";
-import type { SidebarView } from "@/components/ide/shared/ide-types";
-import {
-  type ViewMode,
-  ViewModeProvider,
-} from "@/components/ide/shared/view-mode";
+import { ViewModeProvider } from "@/components/ide/shared/view-mode";
 import { Sidebar } from "@/components/ide/sidebar/sidebar";
 import { SourceControlView } from "@/components/ide/sidebar/source-control-view";
 import { TerminalPanel } from "@/components/ide/terminal/terminal-panel";
-import { navItems } from "@/consts/nav-items";
 import { useEditorGroups } from "@/hooks/use-editor-groups";
 import { useIdeKeyboardShortcuts } from "@/hooks/use-ide-keyboard-shortcuts";
-import { useIsMobile } from "@/hooks/use-is-mobile";
+import { useIdeLayoutState } from "@/hooks/use-ide-layout-state";
 import type { GitCommitItem } from "@/lib/github";
-import { getBreadcrumbPath } from "@/lib/ide/breadcrumb";
 
 interface IdeLayoutProps {
   children: React.ReactNode;
@@ -34,18 +37,9 @@ interface IdeLayoutProps {
 
 export function IdeLayout({ children, commits = [] }: IdeLayoutProps) {
   const pathname = usePathname();
-  const isMobile = useIsMobile();
   const [embed] = useQueryState("embed", parseAsBoolean.withDefault(false));
   const isEmbed = embed;
-  const [sidebarOpen, setSidebarOpen] = useState(true);
-  const [sidebarView, setSidebarView] = useState<SidebarView>("explorer");
-  const [mobileSidebarView, setMobileSidebarView] =
-    useState<SidebarView | null>(null);
-  const [terminalOpen, setTerminalOpen] = useState(false);
-  const [commandOpen, setCommandOpen] = useState(false);
-  const [isFullscreen, setIsFullscreen] = useState(false);
-  const [pageTitle, setPageTitle] = useState("");
-  const [viewMode, setViewMode] = useState<ViewMode>("preview");
+  const contentRef = useRef<HTMLDivElement>(null);
 
   const {
     activeGroupIndex,
@@ -66,71 +60,37 @@ export function IdeLayout({ children, commits = [] }: IdeLayoutProps) {
     splitRight,
   } = useEditorGroups(pathname);
 
-  const contentRef = useRef<HTMLDivElement>(null);
-  const mainRef = useRef<HTMLElement>(null);
+  const layoutState = useIdeLayoutState(activeHref);
 
-  const toggleSidebar = useCallback(() => {
-    setSidebarOpen((prev) => !prev);
-  }, []);
-
-  const toggleTerminal = useCallback(() => {
-    setTerminalOpen((prev) => !prev);
-  }, []);
-
-  const openMobileExplorer = useCallback(() => {
-    setMobileSidebarView("explorer");
-  }, []);
-
-  const openMobileSourceControl = useCallback(() => {
-    setMobileSidebarView("sourceControl");
-  }, []);
+  const {
+    commandOpen,
+    copyContent,
+    focusSourceControl,
+    mainRef,
+    mobileSidebarView,
+    openMobileExplorer,
+    openMobileSourceControl,
+    pageTitle,
+    setCommandOpen,
+    setMobileSidebarView,
+    setSidebarView,
+    setTerminalOpen,
+    setViewMode,
+    sidebarOpen,
+    sidebarView,
+    terminalOpen,
+    toggleFullscreen,
+    toggleSidebar,
+    toggleTerminal,
+    viewMode,
+    isFullscreen,
+  } = layoutState;
 
   useIdeKeyboardShortcuts({
     contentRef,
     onToggleSidebar: toggleSidebar,
     onToggleTerminal: toggleTerminal,
   });
-
-  useEffect(() => {
-    setPageTitle(typeof document !== "undefined" ? document.title : "");
-  }, []);
-
-  useEffect(() => {
-    setMobileSidebarView(null);
-  }, [pathname]);
-
-  useEffect(() => {
-    const onFullscreenChange = () => {
-      setIsFullscreen(!!document.fullscreenElement);
-    };
-    document.addEventListener("fullscreenchange", onFullscreenChange);
-    return () =>
-      document.removeEventListener("fullscreenchange", onFullscreenChange);
-  }, []);
-
-  const toggleFullscreen = useCallback(async () => {
-    try {
-      if (document.fullscreenElement) {
-        await document.exitFullscreen();
-      } else {
-        await document.documentElement.requestFullscreen();
-      }
-    } catch {
-      // Fullscreen not supported or denied
-    }
-  }, []);
-
-  const copyContent = useCallback(() => {
-    const path = activeHref ?? pathname;
-    const title =
-      pageTitle || (typeof document !== "undefined" ? document.title : "");
-    const breadcrumb = getBreadcrumbPath(path, navItems);
-    const mainText = mainRef.current?.innerText ?? "";
-    const formatted = [title, breadcrumb, mainText]
-      .filter(Boolean)
-      .join("\n\n");
-    void navigator.clipboard.writeText(formatted);
-  }, [pathname, pageTitle, activeHref]);
 
   const viewModeValue = useMemo(() => ({ viewMode, setViewMode }), [viewMode]);
 
@@ -223,14 +183,7 @@ export function IdeLayout({ children, commits = [] }: IdeLayoutProps) {
           />
 
           <StatusBar
-            onFocusSourceControl={() => {
-              if (isMobile) {
-                setMobileSidebarView("sourceControl");
-              } else {
-                setSidebarView("sourceControl");
-                setSidebarOpen(true);
-              }
-            }}
+            onFocusSourceControl={focusSourceControl}
             onToggleTerminal={toggleTerminal}
             pathname={pathname}
             terminalOpen={terminalOpen}
